feat(advertisements): auth to delete

+++ src/actions/advertisements-actions.js
@@ -1,7 +1,7 @@
 const { makeAdvertisement } = require('../entities');
 const { advertisementDb } = require('../db');
 
-function buildAdvertisementsActions() {
+function buildAdvertisementsActions({ ForbiddenError }) {
   function findAll() {
     return advertisementDb.findAll();
   }
@@ -15,8 +15,23 @@ function buildAdvertisementsActions() {
     return advertisementDb.insert(newAdvertisement);
   }
 
-  function deleteById({ id }) {
-    return advertisementDb.deleteById({ id });
+  function deleteById({ id, userId }) {
+    return advertisementDb.findById({ id })
+      .then((targetAdvertisement) => {
+        if (targetAdvertisement === null) {
+          throw new Error('Обявление отсутствует');
+        }
+        return targetAdvertisement;
+      })
+      .then((targetAdvertisement) => {
+        if (
+          targetAdvertisement.userId === null
+          || targetAdvertisement.userId.toString() !== userId
+        ) {
+          throw new ForbiddenError('Вы не являетесь автором данного объявления');
+        }
+        return advertisementDb.deleteById({ id });
+      });
   }
 
   return {
